Create MarkdownDemo.md
Create a new Markdown file in the repository called MarkdownDemo.md. Make it demonstrate every piece of Markdown formatting in as few lines of output as possible.
Create a profile README. Make sure that you link to it from inside MarkdownDemo.md
Create a new repository on GitHub and add a file ThingsIDontUnderstand.md to it. Make sure the respository is public access and that you link to it from inside MarkdownDemo.md. Include in ThingsIDontUnderstand.md everything that you don't understand about Git and GitHub and would like to spend more course time on (this may be the empty set).